requirements.txt
pandas>=1.3.0
plotly>=5.0.0
jinja2>=3.0.0
pyyaml>=6.0
numpy>=1.20.0

